Refactor package.json and tsconfig.json generation into their own files
